requirements.txt
websockets==15.0.1
aiohttp==3.11.16
requests==2.32.3
beautifulsoup4==4.13.3
mysql-connector-python==9.2.0
python-dotenv==1.1.0
